feat: 결과데이터 유틸 함수 분리 및 결과 페이지 메인 타입 설정 #16 #29
